+++ src/repo.js
@@ -1,12 +1,12 @@
 const { getOctokit } = require('@actions/github');
-const { debug, info, getInput, setOutput } = require('@actions/core');
+const { debug, info, getInput, setOutput, isDebug } = require('@actions/core');
 const my_token = getInput('my_token', { require: true });
 const octokit = new getOctokit(my_token);
 const { pluck, zip, unzip, reject } = require('underscore');
 const { join } = require('path');
 const { writeFileSync } = require('fs');
 
-let getAll = async function (user, page = 10, isDebug = false) {
+let getAll = async function (user, page = 10) {
   var repo_list = [];
   for (let i = 1; i < parseInt(page); i++) {
     try {
@@ -22,7 +22,7 @@ let getAll = async function (user, page = 10, isDebug = false) {
   }
   var repo_info = join('.repo_list', 'repo-info.json');
   debug(`repo-info: ${repo_info}`);
-  if (isDebug) writeFileSync(repo_info, JSON.stringify(repo_list, null, 2), 'utf-8');
+  if (isDebug()) writeFileSync(repo_info, JSON.stringify(repo_list, null, 2), 'utf-8');
   repo_list = reject(repo_list, item => item.owner.login != user);
   var repo_list_name = pluck(repo_list, 'name');
   var repo_list_private = pluck(repo_list, 'private');